Support eip-8
