+++ src/main/scala/io/iohk/ethereum/network/rlpx/AuthInitiateMessage.scala
@@ -1,4 +1,4 @@
-package io.iohk.ethereum.network
+package io.iohk.ethereum.network.rlpx
 
 import akka.util.ByteString
 import io.iohk.ethereum.crypto._
@@ -39,5 +39,4 @@ case class AuthInitiateMessage(
     nonce ++
     ByteString(if (knownPeer) 1.toByte else 0.toByte)
   }
-
 }
